Attached tests to orders, made export work correctly, made edits to get all of that to work

+++ INTEX/Models/Test.cs
@@ -23,7 +23,7 @@ namespace INTEX.Models
         public int HOURSTOCOMPLETE { get; set; }
         [Required]
         [Display(Name = "Base Price")]
-        public double BASEPRICE { get; set; }
+        public decimal BASEPRICE { get; set; }
         [Required]
         [Display(Name = "Number Of Days")]
         public int NUMBEROFDAYS { get; set; }
